refactor(agent): extract session sink setup from run-worker
